requirements.txt
fuzzywuzzy
GitPython
python_magic
